Update producer and consumer code
