--- Assignment3/consumer.h
@@ -1,11 +1,0 @@
-#ifndef CONSUME_H
-#define CONSUME_H
-
-#include <stdio.h>
-#include <unistd.h>
-#include "macros.h"
-#include "bbuff.h"
-
-void consume(void);
-
-#endif
