--- Assignment3/producer.h
@@ -1,12 +1,0 @@
-#ifndef PRODUCER_H
-#define PRODUCER_H
-
-#include <stdio.h>
-#include <time.h>
-#include <unistd.h>
-#include "macros.h"
-#include "bbuff.h"
-
-void produce(voidPtr arg);
-
-#endif
